+++ build.gradle
@@ -105,8 +105,8 @@ dependencies {
     provided "com.fasterxml.jackson.core:jackson-core:2.9.3"
     provided "org.jetbrains.kotlin:kotlin-stdlib-jdk8:$kotlin_version"
     provided "org.jetbrains.kotlin:kotlin-reflect:$kotlin_version"
-    provided 'com.github.falsehonesty:asmhelper:b9f48209ad'
-    provided 'com.github.chattriggers:rhino:3b4c66c252'
+    provided 'com.github.falsehonesty:asmhelper:js-support-SNAPSHOT'
+    provided 'com.github.chattriggers:rhino:bf177b67bb'
 }
 
 jar {
