+++ src/main/kotlin/com/chattriggers/ctjs/engine/module/ModuleMetadata.kt
@@ -6,6 +6,7 @@ data class ModuleMetadata(
     val name: String? = null,
     val version: String? = null,
     var entry: String? = null,
+    val asmEntry: String? = null,
     val tags: ArrayList<String>? = null,
     val pictureLink: String? = null,
     val creator: String? = null,
