+++ src/main/kotlin/com/chattriggers/ctjs/launch/plugin/CTJSTransformer.kt
@@ -1,5 +1,6 @@
 package com.chattriggers.ctjs.launch.plugin
 
+import com.chattriggers.ctjs.engine.module.ModuleManager
 import me.falsehonesty.asmhelper.BaseClassTransformer
 import net.minecraft.launchwrapper.LaunchClassLoader
 import kotlin.properties.Delegates
@@ -9,29 +10,42 @@ class CTJSTransformer : BaseClassTransformer() {
         super.setup(classLoader)
 
         classLoader.addTransformerExclusion("ct.") // for proguard builds
-        classLoader.addTransformerExclusion("com.chattriggers.ctjs.launch.plugin.")
+        classLoader.addTransformerExclusion("file__") // for rhino generated classes
+        classLoader.addTransformerExclusion("com.chattriggers.ctjs.")
+        classLoader.addTransformerExclusion("com.google.gson.")
+        classLoader.addTransformerExclusion("org.mozilla.javascript")
+        classLoader.addTransformerExclusion("org.mozilla.classfile")
+        classLoader.addTransformerExclusion("com.fasterxml.jackson.core.Version")
+        classLoader.addTransformerExclusion("me.falsehonesty.asmhelper.")
     }
 
     override fun makeTransformers() {
-        injectCrashReport()
-        injectEntityPlayer()
-        injectGuiMainMenu()
-        injectMinecraft()
-        injectScreenshotHelper()
-        injectNetHandlerPlayClient()
-        injectRenderManager()
-        makeGuiScreenInjections()
-
         try {
-            println("Trying to find frames+")
-            Class.forName("io.framesplus.util.NativeUtil")
-            println("Found frames+")
-            HAS_FRAMES_PLUS = true
-        } catch (e: ClassNotFoundException) {
-            // If Frames+ is present, this injection causes MC
-            // to crash
-            println("Did not find frames+")
-            injectEffectRenderer()
+            injectCrashReport()
+            injectEntityPlayer()
+            injectGuiMainMenu()
+            injectMinecraft()
+            injectScreenshotHelper()
+            injectNetHandlerPlayClient()
+            injectRenderManager()
+            makeGuiScreenInjections()
+
+            try {
+                println("Trying to find frames+")
+                Class.forName("io.framesplus.util.NativeUtil")
+                println("Found frames+")
+                HAS_FRAMES_PLUS = true
+            } catch (e: Exception) {
+                // If Frames+ is present, this injection causes MC
+                // to crash
+                println("Did not find frames+")
+                injectEffectRenderer()
+            }
+
+            ModuleManager.setup()
+            ModuleManager.asmPass()
+        } catch (e: Throwable) {
+            e.printStackTrace()
         }
     }
